Fix ASSEMBLY line-5 import, PPH formula, date prefill; bundle WIP
Verified bug fixes (live-tested against info/EFF MAY V6 (1).xlsx):
- Entry Excel import now maps ASSEMBLY line 5 -> ASSEMBLY SMALL (was BIG),
  via shared SectionMetrics.applyAssemblyLine() also reused by
  SplitEntryService.normalizeSection so both import paths agree.
- Actual PPH now uses DL/MP (Output/MP/WT) to match the Excel "PPH" column
  (was Output/DLI/WT).
- Date inputs prefill correctly via spring.mvc.format.date=yyyy-MM-dd.
- Import parse errors now surface to the user (flash key error -> importError).
- Updated EfficiencyCalculatorServiceTest (PPH); added missing
  LineAssignmentService @Mock to SplitEntryServiceTest.

Also bundles prior in-progress work already present in the working tree:
VOC module (entities/repositories/service/controller, V12/V13 migrations,
templates), EFF/salary refactor, RBAC, line scoping, and swaps the
info/ EFF reference workbook (APR -> MAY).

+++ src/main/java/thienloc/manage/dto/SalaryReportDto.java
@@ -17,7 +17,7 @@ public class SalaryReportDto {
         private String line;
         private double sixSPercent;        // 6S pass rate; no record = 100%
         private double reprocessPercent;   // reprocess pass rate (100 - defect); no record = 100%
-        private double effectivePct;       // max(0, sixS - reprocess defect rate)
+        private double effectivePct;       // 6S × (1 − reprocess defect rate), as % (Excel G1×(1−G3))
         private int newStyleCount;         // number of distinct styles
         private long newStyleIncentive;    // sum(qty) × 30,000
         private List<String> gradeLabels;  // ["AA","A","B","C","D","E","LL1","LL2","LL3"] for SEW
@@ -29,13 +29,13 @@ public class SalaryReportDto {
     public static class DayRow {
         private LocalDate date;
         private double targetQuota;  // planned daily quota (from DailyProductionDto.target)
-        private int targetMp;        // standard MP from MasterDb.{section}Mp
+        private double targetMp;     // standard MP resolved by EFF calc (Excel sheet S col C)
         private int mp;              // actual MP
         private double wt;
         private double output;       // actual produced quantity
         private String sec;          // e.g. "SEW10"
         private double effSalary;    // from DailyProductionDto
         private double baseRate;     // floor-looked-up from EffIncentiveRate
-        private long[] gradeAmounts; // baseRate × multiplier[i] × effectivePct / 100 (9 elements)
+        private long[] gradeAmounts; // ROUNDUP(baseRate × mult[i] × effectiveFraction + newStyle/days, −2)
     }
 }
